+++ packages/to-json-schema/src/index.ts
@@ -1,5 +1,5 @@
 export type { JSONSchema7 } from 'json-schema';
-export type { ConversionConfig } from './type.ts';
+export type * from './type.ts';
 
 export * from './functions/index.ts';
 export * from './storages/index.ts';